# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/ambulance": {
      "get": {
        "operationId": "AmbulanceController_findAll",
        "summary": "Get all ambulance requests",
        "parameters": [
          {
            "name": "patientId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "Pending",
                "Dispatched",
                "En Route",
                "Picked Up",
                "At Hospital",
                "Completed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of ambulance requests"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "AmbulanceController_create",
        "summary": "Create an ambulance request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAmbulanceRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request creation result (check success field)"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/ambulance/stats/overview": {
      "get": {
        "operationId": "AmbulanceController_getStats",
        "summary": "Get ambulance statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ambulance statistics retrieved"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/ambulance/patient/{patientId}": {
      "get": {
        "operationId": "AmbulanceController_findByPatient",
        "summary": "Get ambulance requests by patient ID",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient requests retrieved"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/ambulance/active": {
      "get": {
        "operationId": "AmbulanceController_getActiveRequests",
        "summary": "Get all active ambulance requests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active requests retrieved"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/ambulance/assigned/{assignedTo}": {
      "get": {
        "operationId": "AmbulanceController_findByAssignedStaff",
        "summary": "Get ambulance requests by assigned staff",
        "parameters": [
          {
            "name": "assignedTo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Staff assigned requests retrieved"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/ambulance/{id}": {
      "get": {
        "operationId": "AmbulanceController_findById",
        "summary": "Get ambulance request by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request details retrieved"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "put": {
        "operationId": "AmbulanceController_update",
        "summary": "Update an ambulance request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAmbulanceRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request updated successfully"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "AmbulanceController_patchUpdate",
        "summary": "Partially update an ambulance request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAmbulanceRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request updated successfully"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "AmbulanceController_delete",
        "summary": "Delete an ambulance request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request deleted successfully"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/ambulance/{id}/dispatch": {
      "patch": {
        "operationId": "AmbulanceController_dispatch",
        "summary": "Dispatch an ambulance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchAmbulanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ambulance dispatched successfully"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/ambulance/{id}/complete": {
      "patch": {
        "operationId": "AmbulanceController_complete",
        "summary": "Mark an ambulance request as completed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request completed successfully"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/ambulance/{id}/status": {
      "patch": {
        "operationId": "AmbulanceController_updateStatus",
        "summary": "Update ambulance request status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status updated successfully"
          }
        },
        "tags": [
          "Ambulance"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "NexCare Hospital Management API",
    "description": "Complete REST API for the NexCare Hospital Administrative Operations Platform. Features role-based access control (RBAC), JWT authentication, and comprehensive modules for patients, appointments, billing, beds, ambulance, inventory, feedback, and system management.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      },
      "x-user-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "User role for RBAC (superuser, administrative_staff, patient, ambulance, doctor, nurse)"
      }
    },
    "schemas": {
...
      "CreateAmbulanceRequestDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "example": "P001",
            "description": "ID of the patient requesting"
          },
          "patientName": {
            "type": "string",
            "example": "John Doe",
            "description": "Name of the patient requesting"
          },
          "pickupLocation": {
            "type": "string",
            "example": "123 Main Street",
            "description": "Pickup location address"
          },
          "contact": {
            "type": "string",
            "example": "+1 (555) 123-4567",
            "description": "Contact phone number"
          },
          "notes": {
            "type": "string",
            "example": "Patient has chest pain",
            "description": "Additional notes"
          }
        },
        "required": [
          "patientId",
          "pickupLocation",
          "contact"
        ]
      },
      "UpdateAmbulanceRequestDto": {
        "type": "object",
        "properties": {
          "pickupLocation": {
            "type": "string",
            "example": "123 Main Street",
            "description": "Pickup location address"
          },
          "contact": {
            "type": "string",
            "example": "+1 (555) 123-4567",
            "description": "Contact phone number"
          },
          "notes": {
            "type": "string",
            "example": "Patient has chest pain",
            "description": "Additional notes"
          },
          "status": {
            "type": "string",
            "enum": [
              "Pending",
              "Dispatched",
              "En Route",
              "Picked Up",
              "At Hospital",
              "Completed"
            ],
            "example": "Pending",
            "description": "Request status"
          },
          "assignedTo": {
            "type": "string",
            "example": "U003",
            "description": "Staff assigned"
          },
          "stepIndex": {
            "type": "number",
            "example": 0,
            "description": "Current step index in transport"
          },
          "completedDate": {
            "type": "string",
            "example": "May 5, 2026",
            "description": "Date the transport was completed"
          },
          "completedTime": {
            "type": "string",
            "example": "10:30 AM",
            "description": "Time the transport was completed"
          }
        }
      },
      "DispatchAmbulanceDto": {
        "type": "object",
        "properties": {
          "assignedTo": {
            "type": "string",
            "example": "U003",
            "description": "ID of staff assigned"
          },
          "vehicleNumber": {
            "type": "string",
            "example": "AMB-101",
            "description": "Vehicle number"
          },
          "driverContact": {
            "type": "string",
            "example": "+1 (555) 999-8888",
            "description": "Driver contact number"
          },
          "estimatedArrival": {
            "type": "string",
            "example": "10:30 AM",
            "description": "Estimated arrival time"
          },
          "dispatchedBy": {
            "type": "string",
            "example": "U002",
            "description": "Staff ID who dispatched"
          }
        },
        "required": [
          "assignedTo",
          "dispatchedBy"
        ]
      },
...
    }
  }
}